debug: Move some logic from ball roller component in to base class and ball manager.
