+++ VisualPinball.Unity/VisualPinball.Unity/Game/BallDebugComponent.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: dfe9af57cee842449d63397ffed037bc
+timeCreated: 1745065405
